Added Automated Setup and Timer Start command
Resolves #2

Automated Setup Command
=======================

To ensure the app is configured with the proper workspace ID, the ENV
var is checked for on app bootstrap. If it's not found, the setup
command is triggered which loads all workspaces available to the current
API key.

**If there are multiple workspaces,** a list is presented and the user
can choose. If there is only one, it's auto-selected. Once a selection
is made, the unique workspace ID is shown to the user and they're
prompted to add this to their environment.

All Clockify usage requires the use of these unique IDs so searching and
listing all workspaces in a human readable way helps bridge this gap.

Running the setup command directly can be useful should the workspace
need to be changed.

Timer Start Command
===================

One of the primary actions done many times a day is starting a timer
with a set description, required project, and optionally a tag. For both
projects and tags, a request is made to Clockify for projects and tags,
filtering down if one was provided. The user is presented with a list if
there are multiple results, or they are prompted for a confirmation to
use the single value returned.

Once a human readable choice is made, we can use the unique ID going
forwards in the request as required.

Project Option
--------------

For this tool, currently, the project is required. It can either be
passed in with `--project|-p` or, if nothing was passed, all projects
will be fetched and the user will be prompted for a choice.

Tag option
----------

If a tag is to be attached to the timer, it can be passed directly and
used in a fuzzy search, or just passing `--tag` will pull down _all_
tags and prompt for a choice. Currently only a single tag is supported
however in the future, multiple tag selections will be possible.

Base Command
============

A base command was required as the `OutputInterface` doesn't have the
`section()` method which is very useful. To satisfy static analysis,
creating a section is wrapped in a function which does a type check.